Feature Progress:
Continued additions to sign up / email verification / registration flow.

Now LoginActivity checks if user exists. If they do then it checks
to see that if they signed up with email / password that they have verified
the email address.

After these checks the Activity then checks to see if the user exists in
the DB. If so it will bring the user to the MainActivity, and if not
it will begin the RegistrationFragment.