Bug Fix:
Bug: request feed was not displaying on initial load.

Cause: FeedFragment was drawing its views at the same time that the view
model was downloading the request data at the same time that the fragment
was creating the adapter. The adapter was being created before the data
was done downloading and thus, no data was being shown until fragment
was recreated.

Fix: Moved downloading of requests into fragment. This way the success
callback can be used to tell the fragment when to draw the recycler view.