Changed Launcher activity to be MainActivity rather than the LoginActivity.
This way a quick check can be made in MainActivity if the user is
authenticated. If so just load up the FeedFragment. If not then go to the
LoginActivity.

With this change user has better load up time once they are signed in
rather than just a good launch time when first signing up.