Bug FIx:
Bug: When going to the SKillsFragment skills you had selected
would not show as checked.

Cause: SkillsFragment was doing ViewModelProviders.of(this).get(MainActivityViewModel::class.java)

which was instantiating a sepertate view model for the fragment itself.
the view model would not be done loading its intial dataset at the
time of the fragment view creation. AND we would not want to use a
seperate view model anyways..

Fix: Changed view model declaration in fragment to

ViewModelProviders.of(activity!!).get(MainActivityViewModel::class.java)