EditProfileFragment up and running. Still has some kinks:
State and Region SearchableSpinner can not have hint updated.
So Still says State and Region rather than the users
currently selected State and Region. Seems to be no way
to update his field. Tried another searchable spinner library and
same issue. May need to revert to a normal Spinner and see if that
alleviates the problem.

Still some issues with photo uploads as well. At first new photo was
not being displayed but fixed that by clearing Glides memory and
disk cache. Need to look into if there is a way to update the
specific image rather than invalidating all of the cache.

Also, loading of images is still very slow. Need to find a way to
compress images.